refactor(web): relocate server-only modules from lib/ to server/
Move qstash, rate-limit, auth, email, supabase, resend-segment,
crawl-processor, crawl-stale, file-stale, crawled-page-files,
crawler-metadata-sql, file-processor and local-storage into server/.
These import node-only/service dependencies and were one accidental
client import away from leaking secrets into a browser bundle. lib/
now holds only client-safe or isomorphic modules.
